refactor: dedicated modules for ubiquitous utils
