+++ src/common.rs
@@ -1,358 +1,4 @@
 //! Common modules and other globally available items
-//!
-//! The common logging system is a simplified version of what is available in
-//! the `log` crate, and all logging functions shall use this common module.
-//!
-//! Some notes on logging:
-//!
-//! * The log messages are composed by
-//!     - the timestamp
-//!     - the application name (see below) in brackets
-//!     - the log level
-//!     - the log message
-//! * The log message in turn has the following form:
-//!   `context: [MSGTYPE] human readable message`
-//!   where
-//!     - the context is usually constructed with two space-separated strings
-//!       indicating the part of the program where a certain message is issued
-//!     - MSGTYPE (in square brackets) consists of two or more alphanumeric
-//!       strings, separated by slashes, whose first two are described below
-//!       and the further ones may depend on the first two
-//!     - the human readable message is an explanation of what happened.
-//!
-//! The first two elements in MSGTYPE indicate in which point of an operation
-//! the event occurs, and the type of event. The first element can be one of:
-//!
-//! * _INIT_ if the event occurs in an initialization phase
-//! * _START_ if the event occurs while starting something, service or process
-//! * _PROC_ if the event occurs while processing or during some activity
-//! * _END_ if the event occurs at the end of a service or process
-//! * _HIST_ is a _trace level only_ message emitted to show history on GUI:
-//!   in this case _MSG_ is sent at the beginning of task execution, and
-//!   _OK_, _FAIL_ or _IND_ are sent at the end (resp. on success,
-//!   failure or _indeterminate_ outcome)
-//! * _BUSY_ is also a _trace level only_ message emitted to allow a GUI to
-//!   show the application status (for instance using an icon in the
-//!   tray area): when there are one or more conditions busy, the second
-//!   element is _YES_, otherwise _NO_
-//! * _PAUSE_ another _trace level only_ message emitted to allow a GUI to
-//!   change application status (for instance using a tray icon) when
-//!   the scheduler is paused: useful because an _internal command_
-//!   based task might pause the scheduler unattendedly
-//!
-//! while the second can be one of:
-//!
-//! * _OK_ for expected outcomes or behaviours
-//! * _FAIL_ for unexpected outcomes or behaviours
-//! * _IND_ for indeterminate outcomes
-//! * _MSG_ if the human-readable part is exclusively informational
-//! * _ERR_ (may be followed by a dash `-` and a code) for errors to be
-//!   notified
-//! * _YES_ (only occurs for _BUSY_ or _PAUSE_ indicators) means: application
-//!   is busy or has been paused
-//! * _NO_ (only occurs for _BUSY_ or _PAUSE_ indicators) means: application
-//!   is _not_ busy or has been resumed
-//!
-//! This should help using the log as a way of communicating to a wrapper
-//! utility the state of the scheduler, and possibily give the opportunity to
-//! organize communication to the user in a friendlier way.
-//!
-//! This module also contains common enums, traits, structs, and functions
-//! shared between items that use the same technology. Shared collections are
-//! organized in modules:
-//!
-//! * `cmditem` for assets common to command based tasks and conditions
-//! * `luaitem` for assets common to Lua based tasks and conditions
-//! * `dbusitem` for assets common to DBus based conditions and events
-//! * `wmiitem` for assets common to WMI based conditions and events
-//! * `wres` for the _whenever_ specific `Result`, that has automations
-//!   for conversion from many other result types
-//!
-//! in order to avoid behaviour discrepancies, and possibly to save some
-//! memory by avoiding unnecessary duplications.
-
-use lazy_static::lazy_static;
-use parking_lot::RwLock;
-
-// the following global flag is exposed here because it looks like there is
-// no actual way to pass anything but a string as payload to the logger, so
-// the common logging function should know whether the logger is initialized
-// to return JSON message and build the JSON payload itself
-lazy_static! {
-    static ref LOGGER_EMITS_JSON: RwLock<bool> = RwLock::new(false);
-}
-
-#[allow(dead_code)]
-/// Module for logging
-///
-/// Exposes (publicly):
-///
-/// * a function to universally log (`log`)
-/// * a logger initialization function (`init`)
-/// * the logging levels: _trace_ < _debug_ < _info_ < _warn_ < _error_,
-///   provided as the `LogType` enumeration
-pub mod logging {
-    use crate::constants::{APP_NAME, ERR_LOGGER_NOT_INITIALIZED};
-    use flexi_logger::{DeferredNow, FileSpec, Logger, style};
-    use log::Record;
-    use log::{debug, error, info, trace, warn};
-    use nu_ansi_term::Style;
-    use serde_json::json;
-    use std::path::PathBuf;
-
-    use super::LOGGER_EMITS_JSON;
-
-    // time stamp format that is used by the provided format functions.
-    const NOW_FMT: &str = "%Y-%m-%dT%H:%M:%S%.3f";
-    const NOW_FMT_FULL: &str = "%Y-%m-%dT%H:%M:%S%.6f";
-
-    // log formatters
-    fn log_format_plain(
-        w: &mut dyn std::io::Write,
-        now: &mut DeferredNow,
-        record: &Record,
-    ) -> Result<(), std::io::Error> {
-        write!(
-            w,
-            "[{}] ({APP_NAME}) {} {}",
-            now.format(NOW_FMT),
-            format_args!("{:5}", record.level()),
-            record.args(),
-        )
-    }
-
-    fn log_format_json(
-        w: &mut dyn std::io::Write,
-        now: &mut DeferredNow,
-        record: &Record,
-    ) -> Result<(), std::io::Error> {
-        let header = json!({
-            "application": APP_NAME,
-            "time": now.format(NOW_FMT_FULL).to_string(),
-            "level": record.level().to_string(),
-        });
-        let payload = record.args();
-        write!(w, "{{\"header\":{header},\"contents\":{payload}}}")
-    }
-
-    fn log_format_colors(
-        w: &mut dyn std::io::Write,
-        now: &mut DeferredNow,
-        record: &Record,
-    ) -> Result<(), std::io::Error> {
-        let level = record.level();
-        let bold = Style::new().bold();
-        let dimmed = Style::new().dimmed();
-        write!(
-            w,
-            "[{}] {} {} {}",
-            format_args!("{}", now.format(NOW_FMT)),
-            dimmed.paint(format!("({APP_NAME})")),
-            style(level).paint(format!("{:5}", level.to_string())),
-            bold.paint(record.args().to_string()),
-        )
-    }
-
-    /// Log levels (from most verbose to least)
-    pub enum LogType {
-        Trace,
-        Debug,
-        Info,
-        Warn,
-        Error,
-    }
-
-    /// Logger initialization: if `filename` is not given, the log will be
-    /// sent to stdout and use color (and the `append` parameter will be
-    /// ignored); otherwise `filename` will be used as path for the log file:
-    /// causes an error if it's not possible to open the log file.
-    pub fn init(
-        level: LogType,
-        filename: Option<String>,
-        append: bool,
-        logcolor: bool, // these three values are mutually
-        logplain: bool, // exclusive by construction of the
-        logjson: bool,  // main `clap` parser
-    ) -> std::io::Result<bool> {
-        let level = match level {
-            LogType::Trace => "trace",
-            LogType::Debug => "debug",
-            LogType::Info => "info",
-            LogType::Warn => "warn",
-            LogType::Error => "error",
-        };
-
-        // the following line is to avoid other crates logging (e.g. `zbus`)
-        // so it can be commented out for debugging purposes and replaced with
-        // the subsequent commented out line. A reminder to documentation:
-        // https://docs.rs/flexi_logger/latest/flexi_logger/struct.LogSpecification.html
-        // FIXME: maybe we can choose the actual configuration string
-        //        automatically according to the current build settings?
-        let logspec = format!("whenever={level}");
-        // let logspec = format!("{level}");
-
-        let mut logger;
-        logger = Logger::try_with_str(logspec);
-        match logger {
-            Ok(l) => {
-                if let Some(fname) = filename {
-                    let log_format = {
-                        if logcolor {
-                            log_format_plain
-                        } else if logplain {
-                            log_format_plain
-                        } else if logjson {
-                            *LOGGER_EMITS_JSON.write() = true;
-                            log_format_json
-                        } else {
-                            log_format_plain
-                        }
-                    };
-                    let mut pb = PathBuf::from(&fname);
-                    if pb.parent().is_none()
-                        || pb.parent().unwrap().to_str().unwrap_or("").is_empty()
-                    {
-                        pb = {
-                            let mut dir = PathBuf::from(".");
-                            dir.push(pb);
-                            dir
-                        }
-                    }
-                    let fspec = FileSpec::try_from(&pb)
-                        .map_err(|e| std::io::Error::other(e.to_string()))?;
-                    logger = Ok(
-                        l.log_to_file(fspec).format_for_files(log_format), // .write_mode(WriteMode::BufferAndFlush)
-                    );
-                    if append {
-                        logger = Ok(logger.unwrap().append());
-                    }
-                } else {
-                    let log_format = {
-                        if logcolor {
-                            log_format_colors
-                        } else if logplain {
-                            log_format_plain
-                        } else if logjson {
-                            *LOGGER_EMITS_JSON.write() = true;
-                            log_format_json
-                        } else {
-                            log_format_colors
-                        }
-                    };
-                    // in json mode to console we also support capture by pipes
-                    // so that wrappers may use stdout to get updates; it also
-                    // redirects the logger's own errors to a black hole in
-                    // order to avoid polluting a wrapper
-                    if logjson {
-                        logger = Ok(l
-                            .format_for_stdout(log_format)
-                            .write_mode(flexi_logger::WriteMode::Direct)
-                            .error_channel(flexi_logger::ErrorChannel::DevNull)
-                            .log_to_stdout());
-                    } else {
-                        logger = Ok(l.format_for_stdout(log_format).log_to_stdout());
-                    }
-                }
-            }
-            _ => {
-                return Err(std::io::Error::new(
-                    std::io::ErrorKind::Unsupported,
-                    ERR_LOGGER_NOT_INITIALIZED,
-                ));
-            }
-        }
-        if let Err(_e) = logger.unwrap().start() {
-            return Err(std::io::Error::new(
-                std::io::ErrorKind::Unsupported,
-                ERR_LOGGER_NOT_INITIALIZED,
-            ));
-        }
-
-        Ok(true)
-    }
-
-    /// Common log function. The parameters are granular in order to achieve
-    /// two benefits: the first is that for most of them a constant can be
-    /// used, thus reducing the possibility of non-conformant log messages
-    /// (which may arise on typos) and, to some extent, the executable size;
-    /// the second is that JSON log messages can be as fine-grained as
-    /// needed. The constants to be used are defined in _constants.rs_, and
-    /// in particular:
-    ///
-    /// * `emitter` is one of the `LOG_EMITTER_...` constants
-    /// * `action` is one of the `LOG_ACTION_...` constants
-    /// * `when` is one of the `LOG_WHEN_...` constants
-    /// * `status` is one of the `LOG_STATUS_...` constants
-    ///
-    /// while non-constant parameters must be defined as follows
-    ///
-    /// * `item` can be a tuple consisting of item _name_ and _id_
-    /// * `message` is the only arbitrary string that can be passed
-    ///
-    /// This allows JSON messages to be easily interpretable by a wrapper
-    /// according to the hints given in the documentation.
-    pub fn log(
-        severity: LogType,
-        emitter: &str,
-        action: &str,
-        item: Option<(&str, i64)>,
-        when: &str,
-        status: &str,
-        message: &str,
-    ) {
-        let payload = if *LOGGER_EMITS_JSON.read() {
-            let context = if let Some((item, item_id)) = item {
-                json!({
-                    "emitter": emitter,
-                    "action": action,
-                    "item": item,
-                    "item_id": item_id,
-                })
-            } else {
-                json!({
-                    "emitter": emitter,
-                    "action": action,
-                    "item": null,
-                    "item_id": null,
-                })
-            };
-            let message_type = json!({
-                "when": when,
-                "status": status,
-            });
-            json!({
-                "context": context,
-                "message_type": message_type,
-                "message": message,
-            })
-            .to_string()
-        } else {
-            let item_repr = if let Some((name, id)) = item {
-                format!(" {name}/{id}")
-            } else {
-                String::new()
-            };
-            format!("{emitter} {action}{item_repr}: [{when}/{status}] {message}")
-        };
-        match severity {
-            LogType::Trace => {
-                trace!("{payload}")
-            }
-            LogType::Debug => {
-                debug!("{payload}")
-            }
-            LogType::Info => {
-                info!("{payload}")
-            }
-            LogType::Warn => {
-                warn!("{payload}")
-            }
-            LogType::Error => {
-                error!("{payload}")
-            }
-        }
-    }
-}
 
 #[allow(dead_code)]
 /// This module helps command based items perform common activities
@@ -360,7 +6,7 @@ pub mod cmditem {
     use std::time::{Duration, SystemTime};
     use subprocess::{Exec, ExitStatus};
 
-    use crate::LogType;
+    use crate::utility::logging::LogType;
     use crate::constants::*;
 
     /// In case of failure, the reason will be one of the provided values
@@ -1186,168 +832,6 @@ pub mod cmditem {
     }
 }
 
-#[cfg(feature = "lua_sync")]
-#[allow(dead_code)]
-/// A module providing named mutexes that can be shared across all threads.
-///
-/// This module implements a global set of named mutexes that can be locked
-/// and released by any thread.
-pub mod named_mutex {
-    // NOTE: originally this module was AI generated, however the poor guy
-    // kept making mistakes of all sorts, from completely reinventing common
-    // (or standard) libraries API, to writing failing tests, and so on. In
-    // the end, however, it made me discover some interesting libraries and
-    // rethink on how to implement the functionality by hand. Now the module
-    // is hand coded (apart from doc comments)
-    use lazy_static::lazy_static;
-    use parking_lot::{Condvar, Mutex};
-    use std::collections::HashMap;
-    use std::sync::Arc;
-    use std::time::Duration;
-
-    // global map of named mutexes
-    #[derive(Debug)]
-    struct SharedLock {
-        busy: Mutex<bool>,
-        notifier: Condvar,
-    }
-
-    #[allow(dead_code)]
-    impl SharedLock {
-        pub fn new_free() -> Self {
-            SharedLock {
-                busy: Mutex::new(false),
-                notifier: Condvar::new(),
-            }
-        }
-
-        // busy by default
-        pub fn new() -> Self {
-            SharedLock {
-                busy: Mutex::new(true),
-                notifier: Condvar::new(),
-            }
-        }
-
-        // this reclaims a named mutex, possibly with a timeout: if able to
-        // capture it, then it changes its busy state to true and returns
-        // true to signal that it succeeded
-        pub fn claim(&self, timeout: Option<Duration>) -> bool {
-            let mut busy = self.busy.lock();
-            if *busy {
-                if let Some(timeout) = timeout {
-                    if self.notifier.wait_for(&mut busy, timeout).timed_out() {
-                        false
-                    } else {
-                        *busy = true;
-                        true
-                    }
-                } else {
-                    self.notifier.wait(&mut busy);
-                    *busy = true;
-                    true
-                }
-            } else {
-                *busy = true;
-                true
-            }
-        }
-
-        // free the mutex and signal the next waiting thread that it can go on;
-        // this fails only if there was nothing to release
-        pub fn free(&self) -> bool {
-            let mut busy = self.busy.lock();
-            if *busy {
-                *busy = false;
-                self.notifier.notify_one();
-                true
-            } else {
-                false
-            }
-        }
-    }
-
-    lazy_static! {
-        static ref NMUTEX_MAP: Arc<Mutex<HashMap<String, Arc<SharedLock>>>> =
-            Arc::new(Mutex::new(HashMap::new()));
-    }
-
-    // add or retrieve a lock
-    fn get_slock(name: &str) -> Arc<SharedLock> {
-        let map = NMUTEX_MAP.clone();
-        let mut map = map.lock();
-        let s1 = map
-            .entry(name.to_string())
-            .or_insert_with(|| Arc::new(SharedLock::new_free()));
-        s1.clone()
-    }
-
-    // the actual library, as per specification
-
-    /// Attempts to acquire and lock a named mutex.
-    ///
-    /// If a mutex with the given `name` doesn't exist, it creates one and
-    /// locks it immediately.     /// If a mutex with the given `name` exists,
-    /// it attempts to lock it.
-    ///
-    /// # Arguments
-    ///
-    /// * `name` - The name of the mutex to lock
-    /// * `timeout` - Maximum time to wait for the lock.
-    ///   - `None`: Wait indefinitely
-    ///   - `Some(duration)`: Wait for the specified duration
-    ///
-    /// # Returns
-    ///
-    /// Returns `true` if the mutex was successfully locked, `false` if the
-    /// timeout was exceeded.
-    ///
-    /// # Examples
-    ///
-    /// ```ignore
-    /// if namedmutex_lock("Mux01", None) {
-    ///     println!("locked!");
-    ///     std::thread::sleep(std::time::Duration::from_millis(500));
-    ///     let _ = namedmutex_release("Mux01");
-    /// }
-    /// ```
-    ///
-    /// ```ignore
-    /// if !namedmutex_lock("Mux01", Some(Duration::from_millis(1000))) {
-    ///     println!("could not lock the mutex");
-    /// }
-    /// ```
-    pub fn namedmutex_lock(name: &str, timeout: Option<Duration>) -> bool {
-        let sl = &mut get_slock(name);
-        sl.claim(timeout)
-    }
-
-    /// Releases a previously locked named mutex.
-    ///
-    /// # Arguments
-    ///
-    /// * `name` - The name of the mutex to release
-    ///
-    /// # Returns
-    ///
-    /// Returns `true` if the mutex was successfully released, `false` if a
-    /// mutex with the specified name was not found or was not locked by the
-    /// current thread.
-    ///
-    /// # Examples
-    ///
-    /// ```ignore
-    /// if namedmutex_lock("Mux01", None) {
-    ///     // ... do some work ...
-    ///     namedmutex_release("Mux01");
-    /// }
-    /// ```
-    pub fn namedmutex_release(name: &str) -> bool {
-        let sl = &mut get_slock(name);
-        sl.free()
-    }
-}
-
 #[allow(dead_code)]
 /// This module provides utilities for Lua based items
 pub mod luaitem {
@@ -1425,7 +909,7 @@ pub mod luaitem {
     // implement synchronization utilities: sleep and named mutexes
     #[cfg(feature = "lua_sync")]
     pub mod sync {
-        use crate::common::named_mutex::*;
+        use crate::utility::named_mutex::*;
         use crate::constants::*;
         use std::thread;
         use std::time::Duration;
@@ -1649,7 +1133,7 @@ pub mod luaitem {
 #[cfg(feature = "dbus")]
 #[allow(dead_code)]
 pub mod dbusitem {
-    use crate::LogType;
+    use crate::utility::logging::LogType;
     use crate::constants::*;
     use cfgmap::CfgValue;
     use regex::Regex;
@@ -2867,7 +2351,7 @@ pub mod dbusitem {
 #[cfg(feature = "wmi")]
 #[allow(dead_code)]
 pub mod wmiitem {
-    use crate::LogType;
+    use crate::utility::logging::LogType;
     use crate::constants::*;
     use regex::Regex;
     use std::collections::HashMap;
@@ -3102,270 +2586,6 @@ pub mod wmiitem {
             log_message,
         )
     }
-}
-
-/// A common result type: catching errors from modules used throughout
-/// the entire code. The corresponding error carries some information
-/// about what went wrong.
-#[allow(dead_code)]
-pub mod wres {
-    use mlua;
-    use notify;
-    use std::{self, fmt, sync::PoisonError};
-
-    use crate::constants::{ERR_FAILED, ERR_LOCK_FAILED};
-
-    /// Types of specific errors
-    #[non_exhaustive]
-    #[derive(Debug, Clone)]
-    pub enum Kind {
-        Forbidden,
-        Unsupported,
-        Unavailable,
-        Unconverted,
-        Unparsed,
-        Busy,
-        Invalid,
-        Failed,
-        Empty,
-        // ...
-        Unknown,
-    }
-
-    impl fmt::Display for Kind {
-        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-            write!(
-                f,
-                "{}",
-                match self {
-                    Kind::Forbidden => "not permitted",
-                    Kind::Unsupported => "not supported",
-                    Kind::Unavailable => "not available",
-                    Kind::Unconverted => "not converted",
-                    Kind::Unparsed => "not parsed",
-                    Kind::Busy => "resource busy",
-                    Kind::Invalid => "invalid",
-                    Kind::Failed => "failed",
-                    Kind::Empty => "empty",
-                    Kind::Unknown => "unknown",
-                }
-            )
-        }
-    }
-
-    /// Describes the origin of the error: if `Native` the error was originated
-    /// natively, otherwise the field is set by another error that is converted
-    /// into `Error` via a dedicated `From` trait implementation.
-    #[non_exhaustive]
-    #[derive(Debug, Clone, PartialEq)]
-    pub enum Origin {
-        Native,
-        Unit,
-        StdIo,
-        Notify,
-        Sync,
-        Lua,
-
-        #[cfg(feature = "dbus")]
-        DBus,
-
-        #[cfg(windows)]
-        #[cfg(feature = "wmi")]
-        Wmi,
-
-        // ...
-        Unknown,
-    }
-
-    impl fmt::Display for Origin {
-        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-            write!(
-                f,
-                "{}",
-                match self {
-                    Origin::Native => "self",
-                    Origin::Unit => "unit",
-                    Origin::StdIo => "io",
-                    Origin::Notify => "fschange",
-                    Origin::Sync => "sync",
-                    Origin::Lua => "lua",
-
-                    #[cfg(feature = "dbus")]
-                    Origin::DBus => "dbus",
-
-                    #[cfg(windows)]
-                    #[cfg(feature = "wmi")]
-                    Origin::Wmi => "wmi",
-
-                    // ...
-                    Origin::Unknown => "unknown",
-                }
-            )
-        }
-    }
-
-    /// The error type that is used throughout the application: implementations
-    /// of the `From` trait are used to implicitly convert from other error
-    /// types, which in turn set the `origin` property.
-    #[derive(Debug, Clone)]
-    pub struct Error {
-        kind: Kind,
-        origin: Origin,
-        message: String, // freeform message: owned in order to avoid lifetime management
-    }
-
-    impl fmt::Display for Error {
-        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-            if self.origin != Origin::Native {
-                write!(f, "{} ({}): {}", self.kind, self.origin, self.message)
-            } else {
-                write!(f, "{}: {}", self.kind, self.message)
-            }
-        }
-    }
-
-    // maybe the most important: From<std::io::Error>
-    impl From<std::io::Error> for Error {
-        fn from(e: std::io::Error) -> Self {
-            Self {
-                kind: match e.kind() {
-                    std::io::ErrorKind::Unsupported => Kind::Unsupported,
-                    std::io::ErrorKind::PermissionDenied => Kind::Forbidden,
-                    std::io::ErrorKind::InvalidData => Kind::Invalid,
-                    std::io::ErrorKind::InvalidInput => Kind::Invalid,
-                    _ => Kind::Unknown,
-                },
-                origin: Origin::StdIo,
-                message: e.to_string(),
-            }
-        }
-    }
-
-    // notify (fschange) errors
-    impl From<notify::Error> for Error {
-        fn from(e: notify::Error) -> Self {
-            Self {
-                kind: Kind::Failed,
-                origin: Origin::Notify,
-                message: e.to_string(),
-            }
-        }
-    }
-
-    // Lua errors
-    impl From<mlua::Error> for Error {
-        fn from(e: mlua::Error) -> Self {
-            Self {
-                kind: Kind::Failed,
-                origin: Origin::Lua,
-                message: e.to_string(),
-            }
-        }
-    }
-
-    // zbus errors
-    #[cfg(feature = "dbus")]
-    impl From<zbus::Error> for Error {
-        fn from(e: zbus::Error) -> Self {
-            Self {
-                kind: Kind::Failed,
-                origin: Origin::DBus,
-                message: e.to_string(),
-            }
-        }
-    }
-
-    // wmi errors
-    #[cfg(windows)]
-    #[cfg(feature = "wmi")]
-    impl From<wmi::WMIError> for Error {
-        fn from(e: wmi::WMIError) -> Self {
-            let kind = match e {
-                wmi::WMIError::ConvertBoolError(_)
-                | wmi::WMIError::ConvertStringError(_)
-                | wmi::WMIError::ConvertLengthError(_)
-                | wmi::WMIError::ConvertDatetimeError(_)
-                | wmi::WMIError::ConvertDurationError(_)
-                | wmi::WMIError::ConvertVariantError(_)
-                | wmi::WMIError::ConvertError(_) => Kind::Unconverted,
-                wmi::WMIError::DeserializeValueError(_)
-                | wmi::WMIError::InvalidDeserializationVariantError(_)
-                | wmi::WMIError::SerdeError(_) => Kind::Invalid,
-                wmi::WMIError::ParseDatetimeError(_)
-                | wmi::WMIError::ParseFloatError(_)
-                | wmi::WMIError::ParseIntError(_) => Kind::Unparsed,
-                wmi::WMIError::UnimplementedArrayItem => Kind::Unavailable,
-                _ => Kind::Failed,
-            };
-            Self {
-                kind,
-                origin: Origin::Wmi,
-                message: e.to_string(),
-            }
-        }
-    }
-
-    // resource locking errors
-    impl<T> From<PoisonError<T>> for Error {
-        fn from(_: PoisonError<T>) -> Self {
-            Self {
-                kind: Kind::Failed,
-                origin: Origin::Sync,
-                message: ERR_LOCK_FAILED.to_owned(),
-            }
-        }
-    }
-
-    // errors based on the unit type
-    impl From<()> for Error {
-        fn from(_: ()) -> Self {
-            Self {
-                kind: Kind::Failed,
-                origin: Origin::Unit,
-                message: ERR_FAILED.to_owned(),
-            }
-        }
-    }
-
-    // implements `Error` and provides access to properties
-    impl Error {
-        // this is used only to natively create an instance of `Error`: only
-        // conversions set the `origin` property to something different
-        pub fn new(kind: Kind, message: &str) -> Self {
-            Self {
-                kind,
-                origin: Origin::Native,
-                message: message.to_string(),
-            }
-        }
-
-        // property access
-        pub fn kind(&self) -> &Kind {
-            &self.kind
-        }
-
-        pub fn origin(&self) -> &Origin {
-            &self.origin
-        }
-
-        pub fn message(&self) -> &str {
-            &self.message
-        }
-    }
-
-    // possible last resort to allow conversions from pointers to errors
-    impl<T: std::error::Error> From<Box<T>> for Error {
-        fn from(e: Box<T>) -> Self {
-            Self {
-                kind: Kind::Unknown,
-                origin: Origin::Unknown,
-                message: e.to_string(),
-            }
-        }
-    }
-
-    /// Specific `Result` type that assumes `wres::Error` as its Err variant
-    pub type Result<T> = std::result::Result<T, Error>;
 }
 
 // end.
