+++ src/condition/bucket_cond.rs
@@ -20,8 +20,8 @@ use cfgmap::CfgMap;
 use regex::Regex;
 
 use super::base::Condition;
-use crate::common::logging::{LogType, log};
-use crate::common::wres::Result;
+use crate::utility::logging::{LogType, log};
+use crate::utility::result::Result;
 use crate::task::registry::TaskRegistry;
 use crate::{cfg_mandatory, constants::*};
