+++ src/condition/command_cond.rs
@@ -29,8 +29,8 @@ use cfgmap::CfgMap;
 
 use super::base::Condition;
 use crate::common::cmditem::*;
-use crate::common::logging::{LogType, log};
-use crate::common::wres::{Error, Kind, Result};
+use crate::utility::logging::{LogType, log};
+use crate::utility::result::{Error, Kind, Result};
 use crate::task::registry::TaskRegistry;
 use crate::{cfg_mandatory, constants::*};
