+++ src/condition/dbus_cond.rs
@@ -22,8 +22,8 @@ use std::str::FromStr;
 
 use super::base::Condition;
 use crate::common::dbusitem::*;
-use crate::common::logging::{LogType, log};
-use crate::common::wres::Result;
+use crate::utility::logging::{LogType, log};
+use crate::utility::result::Result;
 use crate::task::registry::TaskRegistry;
 use crate::{cfg_mandatory, constants::*};
