+++ src/condition/interval_cond.rs
@@ -11,8 +11,8 @@ use std::time::{Duration, Instant};
 use cfgmap::CfgMap;
 
 use super::base::Condition;
-use crate::common::logging::{LogType, log};
-use crate::common::wres::Result;
+use crate::utility::logging::{LogType, log};
+use crate::utility::result::Result;
 use crate::task::registry::TaskRegistry;
 use crate::{cfg_mandatory, constants::*};
