+++ src/condition/time_cond.rs
@@ -25,8 +25,8 @@ use cfgmap::CfgMap;
 use chrono::prelude::*;
 
 use super::base::Condition;
-use crate::common::logging::{LogType, log};
-use crate::common::wres::{Error, Kind, Result};
+use crate::utility::logging::{LogType, log};
+use crate::utility::result::{Error, Kind, Result};
 use crate::task::registry::TaskRegistry;
 use crate::{cfg_mandatory, constants::*};
