+++ src/event/dbus_event.rs
@@ -22,8 +22,8 @@ use zbus;
 
 use super::base::Event;
 use crate::common::dbusitem::*;
-use crate::common::logging::{LogType, log};
-use crate::common::wres::Result;
+use crate::utility::logging::{LogType, log};
+use crate::utility::result::Result;
 use crate::condition::bucket_cond::ExecutionBucket;
 use crate::condition::registry::ConditionRegistry;
 use crate::{cfg_mandatory, constants::*};
