+++ src/event/wmi_event.rs
@@ -14,8 +14,8 @@ use async_trait::async_trait;
 use wmi::WMIConnection;
 
 use super::base::Event;
-use crate::common::logging::{LogType, log};
-use crate::common::wres::Result;
+use crate::utility::logging::{LogType, log};
+use crate::utility::result::Result;
 use crate::condition::bucket_cond::ExecutionBucket;
 use crate::condition::registry::ConditionRegistry;
 use crate::{cfg_mandatory, constants::*};
