+++ src/task/command_task.rs
@@ -38,8 +38,8 @@ use cfgmap::CfgMap;
 // we implement the Task trait here in order to enqueue tasks
 use super::base::Task;
 use crate::common::cmditem::*;
-use crate::common::logging::{LogType, log};
-use crate::common::wres::{Error, Kind, Result};
+use crate::utility::logging::{LogType, log};
+use crate::utility::result::{Error, Kind, Result};
 use crate::{cfg_mandatory, constants::*};
 
 use crate::cfghelp::*;
