+++ src/task/lua_task.rs
@@ -19,9 +19,9 @@ use mlua;
 
 // we implement the Task trait here in order to enqueue tasks
 use super::base::Task;
-use crate::common::logging::{LogType, log};
+use crate::utility::logging::{LogType, log};
 use crate::common::luaitem::*;
-use crate::common::wres::{Error, Kind, Result};
+use crate::utility::result::{Error, Kind, Result};
 use crate::{cfg_mandatory, constants::*};
 
 use crate::cfghelp::*;
